Restructure settings files, rename, extract styles
