+++ publisher/src/components/Settings/Settings.styles.tsx
@@ -1,0 +1,59 @@
+// Recidiviz - a data platform for criminal justice reform
+// Copyright (C) 2022 Recidiviz, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+// =============================================================================
+
+import styled from "styled-components/macro";
+
+import { typography } from "../GlobalStyles";
+
+export const SettingsContainer = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  padding: 39px 24px 0 24px;
+  position: fixed;
+  overflow-y: scroll;
+
+  @media only screen and (max-width: 1050px) {
+    width: unset;
+    flex-direction: column;
+  }
+`;
+
+export const SettingsFormUploadedFilesWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  flex: 3 1 auto;
+`;
+
+export const SettingsTitle = styled.div`
+  ${typography.sizeCSS.headline}
+  display: flex;
+  flex: 1 1 auto;
+`;
+
+export const SettingsFormPanel = styled.div``;
+
+export const InputWrapper = styled.div`
+  display: flex;
+  gap: 10px;
+
+  div {
+    width: 100%;
+  }
+`;
